v6.27
+ DietPi-PREP | When we remove non-required systemd units, remove them only if they do not belong to any kept APT package, else mask. This prevents future service reinstall and especially reactivation on package upgrades, in case of ARMbian services harmful in combination with DietPi-RAMlog and tmpfs + ARMbian zRam double mount:
```
Setting up linux-buster-root-next-tinkerboard (5.95) ...
Created symlink /etc/systemd/system/sysinit.target.wants/armbian-ramlog.service → /lib/systemd/system/armbian-ramlog.service.
```

+++ PREP_SYSTEM_FOR_DIETPI.sh
@@ -1024,14 +1024,15 @@ _EOF_
 		# - Stop, disable and remove not required 3rd party services
 		local aservices=(
 
-			# - ARMbian
+			# ARMbian
 			'firstrun'
 			'resize2fs'
 			'log2ram'
 			'armbian*'
-			# - Meveric
+			'tinker-bluetooth'
+			# Meveric
 			'cpu_governor'
-			# - RPi
+			# RPi
 			'sshswitch'
 
 		)
@@ -1045,7 +1046,15 @@ _EOF_
 
 				[[ -e $j ]] || continue
 				[[ -f $j ]] && systemctl disable --now ${j##*/}
-				rm -R $j
+				# Remove if not attached to any APT package, else mask
+				if dpkg -S $j &> /dev/null; then
+
+					systemctl mask ${j##*/}
+
+				else
+					rm -R $j
+
+				fi
 
 			done
